refactor: InviteModal 위치 수정
