+++ src/pages/WholeCalendar/index.tsx
@@ -6,7 +6,7 @@ import { useEffect, useState } from 'react';
 import { useLocation, useParams } from 'react-router-dom';
 import { useRecoilState } from 'recoil';
 import { Calendar } from '@/components/@common';
-import InviteModal from '@/components/@common/Modal/UserConfigModal/InviteModal';
+import InviteModal from '@/components/@common/Modal/InviteModal';
 
 const WholeCalendar = () => {
   const { groupId } = useParams();
